Handheld Companion 0.34.4.0
Handheld Companion 0.33.0.0
New
Added a 3D Gamepad Overlay hotkey for easier controller visualization and IMU calibration.
Added a Gamepad preview to the Device page to assist with IMU calibration.
Device-specific settings are now stored in dedicated JSON configuration files.
Device configuration loading has been redesigned to automatically inherit settings from parent device families when needed.
Device support
Added hardware specification support (cTDP, nTDP, CPU/GPU clocks, TjMax) to improve per-device power management.
Added TCTL (CPU temperature limit) support.
Corrected IMU mapping for the GPD Win Mini and OneXPlayer X1 AMD (thanks to @AnchovyGT22).
Improvements
Improved the Device and Sensor interfaces.
Restored ViGEm support for Xbox 360 and DualShock 4 emulation.
Updated Ryzen SMU support.
Added IMU calibration reset functionality.
Improved detection of Legion wireless controllers.
Fixes
Fixed axis remapping cache invalidation.
Fixed several settings initialization issues.
Fixed controller enumeration with updated XInput libraries.
General code cleanup and stability improvements.
Handheld Companion 0.34.0.0
Improvements
Various UI refinements.
Updated USBIP to 0.9.7.8.
Reduced unnecessary virtual controller updates by intelligently filtering insignificant input changes, improving performance.
Handheld Companion 0.34.1.x
Improvements
Various UI refinements.
Improved controller attachment by using the detected XInput slot.
Added support for cHTC temperature and APU skin temperature controls (thanks to @JamesCJ60).
Sensor system
Completely refactored sensor management:
New Auto sensor selection mode.
Cleaner sensor lifecycle management.
Improved separation between UI and sensor logic.
Better responsiveness and maintainability.
Stability
Improved application locking logic to reduce UI freezes.
Restored CPU temperature limits after resuming from sleep.
Updated SDL3-CS to 3.4.12.1.
Improved controller disconnection detection.
Improved virtual controller startup reliability.
Updated libVIIPER.
Fixes
Fixed several JSON configuration mismatches.
Fixed random crashes related to AutomationElement access.
Fixed ViGEm implementation.
Fixed compatibility issues between USBIP 0.9.7.8 and VIIPER.
Fixed Library and Performance pages remaining visible when their components were disabled.
Handheld Companion 0.34.2.x
Thermal controls
CPU and skin temperature limits are now configured independently on AMD devices.
Temperature limits are automatically restored after sleep.
Improved AMD thermal command handling and localization.
Installer
Added an option to create a Windows Restore Point before installation (enabled by default).
Improvements
Improved TDP and power profile management to avoid unnecessary resets and improve profile transitions.
Increased internal synchronization timeouts to improve reliability under heavy system load.
Fixes
Fixed an IndexOutOfRangeException that could occur in Quick Keyboard with dead keys.
Various UI refinements.
Handheld Companion 0.34.3.x
New
Added configurable Hotkey Rumble Strength, allowing users to customize vibration intensity.
Improvements
Continued work to eliminate UI lockups.
Updated:
SDL3-CS 3.4.12.3
Sentry 6.7.0
System.* dependencies
Fixes
Fixed a regression affecting GPD Win 5 gyro settings.
Added support for Inno Setup 7.
Fixed controller reconnection after disconnect.
Added horizontal mouse wheel scrolling support.
Improved startup sequence to prevent long UI stalls.
Fixed frame limiter maximum frequency detection.
Migrated additional UI operations to asynchronous execution.
Adjusted Serilog logging thresholds.
Handheld Companion 0.34.4.0
Improvements
Temperature limit management is now integrated directly into the device lifecycle.
CPU and skin temperature limits are automatically reapplied whenever the system switches between battery and AC power.
Further internal reliability and architecture improvements.

+++ HandheldCompanion/Misc/FanProfile.cs
@@ -65,7 +65,7 @@ namespace HandheldCompanion.Misc
             temp = Math.Clamp(temp, 0.0, 100.0);
 
             // Trip latch if we hit or exceed Tjmax
-            double tjmax = IDevice.GetCurrent().Tjmax;
+            uint tjmax = IDevice.GetCurrent().Tjmax;
             if (temp >= tjmax)
             {
                 if (!tjLatch)
